pipelines/jobs/ip2geo/ip2geo_requirements.txt
python-dotenv==1.1.0
requests==2.32.3
urllib3==2.4.0
